Add more spacing in the back end views (see #7299)
Description
-----------

This is a little experiment to improve the back end UX.

So far, headers and rows always had the same height:

<img width="826" alt="" src="https://github.com/contao/contao/assets/1192057/26dcb9b2-d0de-42bf-a4b6-0c1e7006b307">

---

This PR puts more emphasize on the rows, because that is what users actually work with:

<img width="826" alt="" src="https://github.com/contao/contao/assets/1192057/d5fc6c6e-54ec-456c-9c1b-4da147c8c216">

---

The effect is especially visible in the article tree:

<img width="826" alt="" src="https://github.com/contao/contao/assets/1192057/ff936b6f-942a-4471-b145-c2ba51862834">

---

But the other views benefit as well:

<img width="826" alt="" src="https://github.com/contao/contao/assets/1192057/21e1771b-c3a1-4f51-8884-1cf1141f0db9">

<img width="828" alt="" src="https://github.com/contao/contao/assets/1192057/7167a7ee-49bd-414f-9984-4c9a6017a322">

<img width="827" alt="" src="https://github.com/contao/contao/assets/1192057/d333cc1a-ae7e-438d-b68f-a0b288bc2193">

<img width="829" alt="" src="https://github.com/contao/contao/assets/1192057/d056a3fd-832b-4567-9d5e-a46fb4f8d75c">

Please check out the branch and see if you like it. Do not just look at the screenshots. 😄

Commits
-------

92dda8af Add more spacing in the back end views
0b4c7154 Adjust the comments style sheet
ae460799 Rebuild the assets